feat: lazy decryption for keyfiles
When performing signing for a given key, gpg will interrogate the
gpg-agent for availability of keygrips of _all_ signing keys, not just
the one that ends up being used for the actual signing.

piv-agent used to pre-emptively decrypted the keyfiles when checking
them against the supplied keygrip. This change causes it to wait until
the actual point of signature before decrypting the keyfiles.

This means that you avoid a pinentry prompt for keyfiles when signing
using a hardware security device.

This change also makes a couple of error handling changes to make
piv-agent more robust when cancelling or failing pinentry prompts.