feat: switch to a supported openpgp library
The OpenPGP support is deprecated in the Go standard library is no
longer supported.

The ProtonMail openpgp fork of the Go standard library has diverged
quite heavily from the its upstream. Particularly problematic for
piv-agent:

* ECDSA keys use a bespoke type rather than the standard library ECDSA
  key types (wtf?), so now we have to do conversions between the two
  types.
* "Public" APIs take private (internal) types in their signature which
  effectively makes them uncallable since the private types cannot be
  instantiated (wtf? wtf!?)

The last point in particular necessitated a fork to add a functional API
for creating a public ECDSA key. Once a similar interface is available
upstream we can switch back.